Updated "Dear ImGui" to v1.80.

+++ clean.sh
@@ -1,12 +1,24 @@
 #!/bin/bash
 
-echo ""
+echo ''
 
 make clean
 
-rm -r CMakeFiles
-rm CMakeCache.txt
-rm "cmake_install.cmake"
+if [ -f 'make.log' ]; then
+	rm 'make.log'
+fi
 
-echo "  Cleaned it all."
-echo ""
+if [ -d 'CMakeFiles' ]; then
+	rm -r 'CMakeFiles'
+fi
+
+if [ -f 'CMakeCache.txt' ]; then
+	rm 'CMakeCache.txt'
+fi
+
+if [ -f 'cmake_install.cmake' ]; then
+	rm 'cmake_install.cmake'
+fi
+
+echo '  Cleaned it all.'
+echo ''
